requirements.txt
torch>=2.1.0
transformers>=4.41.0
peft>=0.10.0
accelerate>=0.30.0
datasets>=2.20.0
bitsandbytes>=0.43.0
gradio>=4.30.0
pyyaml
python-dotenv
nltk
sacrebleu
sentence-transformers